Global: Fix some type comparison/shadow warnings.

+++ Core/ELF/ElfReader.cpp
@@ -29,7 +29,7 @@ const char *ElfReader::GetSectionName(int section) const {
 		return nullptr;
 
 	int nameOffset = sections[section].sh_name;
-	if (nameOffset < 0 || nameOffset >= size_) {
+	if (nameOffset < 0 || (size_t)nameOffset >= size_) {
 		ERROR_LOG(LOADER, "ELF: Bad name offset %d in section %d (max = %d)", nameOffset, section, (int)size_);
 		return nullptr;
 	}
